+++ frontend/package.json
@@ -33,7 +33,6 @@
     "clsx": "^2.1.0",
     "color": "^4.0.1",
     "construct-style-sheets-polyfill": "^3.1.0",
-    "copy-webpack-plugin": "^12.0.2",
     "css-loader": "^6.3.0",
     "date-fns": "^3.6.0",
     "del-cli": "^4.0.1",
@@ -46,8 +45,6 @@
     "eslint-plugin-import-x": "^0.4.1",
     "eslint-plugin-lit": "^1.11.0",
     "eslint-plugin-wc": "^2.0.4",
-    "eslint-webpack-plugin": "^4.1.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.6",
     "fuse.js": "^6.5.3",
     "glob": "^8.1.0",
     "highlight.js": "^11.8.0",
@@ -76,14 +73,9 @@
     "slugify": "^1.6.6",
     "style-loader": "^3.3.0",
     "tailwindcss": "^3.4.1",
-    "terser-webpack-plugin": "^5.3.10",
-    "ts-loader": "^9.2.6",
-    "tsconfig-paths-webpack-plugin": "^4.1.0",
     "typescript": "^5.3.3",
     "update-dotenv": "^1.1.1",
     "url-pattern": "^1.0.3",
-    "webpack": "^5.88.0",
-    "webpack-cli": "^4.8.0",
     "webpack-merge": "^5.8.0",
     "yaml": "^2.0.0-11",
     "zod": "^3.23.8"
@@ -93,10 +85,10 @@
     "postinstall": "del-cli ./node_modules/.cache/webpack",
     "test": "web-test-runner \"src/**/*.test.{ts,js}\"",
     "prebuild": "del-cli ./dist",
-    "build": "webpack --config webpack.prod.js",
-    "build-dev": "webpack --mode development",
-    "build:analyze": "BUNDLE_ANALYZER=true webpack --config webpack.prod.js",
-    "start": "webpack serve --mode=development --config webpack.dev.js",
+    "build": "rspack build --config rspack.prod.js",
+    "build-dev": "rspack build --mode development",
+    "build:analyze": "BUNDLE_ANALYZER=true webpack --config rspack.prod.js",
+    "start": "rspack serve --mode=development --config rspack.dev.js",
     "serve": "node scripts/serve.js",
     "lint": "eslint --fix .",
     "lint:check": "eslint .",
@@ -108,18 +100,21 @@
   },
   "devDependencies": {
     "@lit/localize-tools": "^0.8.0",
+    "@rspack/cli": "^1.1.8",
+    "@rspack/core": "^1.1.8",
     "@types/webpack-bundle-analyzer": "^4.7.0",
     "@web/dev-server-esbuild": "^0.3.3",
     "@web/dev-server-import-maps": "^0.2.0",
     "@web/dev-server-rollup": "^0.6.1",
+    "eslint-rspack-plugin": "^4.2.1",
     "husky": "^8.0.3",
     "lint-staged": "^13.1.0",
     "prettier-plugin-tailwindcss": "^0.5.12",
     "rollup-plugin-typescript-paths": "^1.4.0",
     "sinon": "^12.0.1",
+    "ts-checker-rspack-plugin": "^1.1.0",
     "ts-lit-plugin": "^2.0.1",
-    "webpack-bundle-analyzer": "^4.10.1",
-    "webpack-dev-server": "^4.15.0"
+    "webpack-bundle-analyzer": "^4.10.1"
   },
   "optionalDependencies": {
     "@esm-bundle/chai": "^4.3.4-fix.0",
