+++ frontend/tsconfig.json
@@ -14,6 +14,7 @@
     "inlineSources": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
+    "isolatedModules": true,
     "plugins": [
       {
         "name": "ts-lit-plugin",
